Fix utility rune module runtime type exports
