+++ packages/components/package.json
@@ -4897,6 +4897,7 @@
     "check:no-cycle-imports": "bun run scripts/check-no-cycle-imports.ts",
     "check:no-inline-match-media": "bun run scripts/check-no-inline-match-media.ts",
     "check:placeholder-docs": "bun run scripts/check-placeholder-docs.ts",
+    "check:svelte-ts-runtime-types": "bun run scripts/check-svelte-ts-runtime-types.ts",
     "check:test-cleanup": "bun run scripts/check-test-cleanup.ts",
     "clean": "rm -rf dist coverage tsconfig.tsbuildinfo lostgradient-cinder-*.tgz cinder-*.tgz tmp fixtures/node-consumer/node_modules fixtures/node-consumer/dist fixtures/sveltekit-consumer/node_modules fixtures/sveltekit-consumer/.svelte-kit fixtures/sveltekit-consumer/build fixtures/manifest-consumer/node_modules fixtures/typescript-consumer/node_modules fixtures/typescript-consumer/src/generated fixtures/examples-consumer/node_modules fixtures/examples-consumer/.svelte-kit fixtures/examples-consumer/build fixtures/examples-consumer/src/generated fixtures/examples-consumer/src/routes/+page.svelte",
     "colors:audit": "bun run scripts/check-component-css-raw-colors.ts",
@@ -4908,7 +4909,7 @@
     "components:promotion-check": "bun run scripts/check-promotion-readiness.ts",
     "exports:check": "bun run scripts/generate-exports.ts --check",
     "exports:generate": "bun run scripts/generate-exports.ts",
-    "lint": "oxlint --type-aware --tsconfig ./tsconfig.json && bun run check:no-cycle-imports && bun run check:no-bare-console-warn && bun run check:no-inline-match-media && bun run check:data-cinder-boolean-attributes && bun run check:test-cleanup && bun run tokens:literals -- --strict",
+    "lint": "oxlint --type-aware --tsconfig ./tsconfig.json && bun run check:no-cycle-imports && bun run check:no-bare-console-warn && bun run check:no-inline-match-media && bun run check:svelte-ts-runtime-types && bun run check:data-cinder-boolean-attributes && bun run check:test-cleanup && bun run tokens:literals -- --strict",
     "lint:fix": "oxlint --fix --type-aware --tsconfig ./tsconfig.json",
     "prepack": "bun run build",
     "package:weight": "bun run scripts/report-package-weight.ts",
