+++ packages/components/src/components/sortable-list/sortable-list.types.ts
@@ -3,7 +3,7 @@ export type {
   SortableAnnouncements,
   SortableItemContext,
   SortableReorderChange,
-} from '../../utilities/sortable-controller.svelte.ts';
+} from '../../utilities/sortable-controller.types.ts';
 /** Props for the SortableList component. */
 export type SortableListProps<Item> = {
   /** The list of items to render. */
@@ -24,15 +24,15 @@ export type SortableListProps<Item> = {
   /** Fires with the full reordered array and change metadata on drop. */
   onreorder: (
     nextItems: Item[],
-    change: import('../../utilities/sortable-controller.svelte.ts').SortableReorderChange,
+    change: import('../../utilities/sortable-controller.types.ts').SortableReorderChange,
   ) => void;
   /** Optional overrides for announcement strings. */
   announcements?: Partial<
-    import('../../utilities/sortable-controller.svelte.ts').SortableAnnouncements
+    import('../../utilities/sortable-controller.types.ts').SortableAnnouncements
   >;
   /** Row content snippet. Receives the item and a per-row context. */
   children: Snippet<
-    [Item, import('../../utilities/sortable-controller.svelte.ts').SortableItemContext]
+    [Item, import('../../utilities/sortable-controller.types.ts').SortableItemContext]
   >;
   /** Accessible name for the list (applied as aria-label on the list root). */
   label?: string;
