+++ packages/components/src/utilities/use-announcer.svelte.ts
@@ -26,32 +26,7 @@
  * ```
  */
 
-export interface AnnouncerOptions {
-  /**
-   * Delay before clearing the message (ms).
-   * This should be long enough for screen readers to read the message.
-   * Default: 1000
-   */
-  clearDelay?: number;
-
-  /**
-   * Debounce delay before announcing (ms).
-   * Use when announcements may fire rapidly (e.g., typing indicators).
-   * Default: 0 (no debounce)
-   */
-  debounceMs?: number;
-}
-
-export interface Announcer {
-  /** The current message to display in the live region */
-  readonly message: string;
-  /** Announce a message to screen readers */
-  announce(text: string): void;
-  /** Clear any pending announcement and reset message */
-  clear(): void;
-  /** Cleanup resources - call via onDestroy() if needed before component unmount */
-  destroy(): void;
-}
+import type { Announcer, AnnouncerOptions } from './use-announcer.types.ts';
 
 /**
  * Create a screen reader announcer for use with ARIA live regions.
