+++ packages/components/src/utilities/use-intersection.svelte.ts
@@ -1,17 +1,8 @@
-import type { Attachment } from 'svelte/attachments';
-
-export type IntersectionCallback = (entry: IntersectionObserverEntry) => void;
-
-export type UseIntersectionOptions = {
-  /** Static root element. If null/undefined, observes against the viewport. */
-  root?: Element | Document | null;
-  /** CSS margin string captured when the attachment mounts. Defaults to "0px". */
-  rootMargin?: string;
-  /** Threshold(s) captured when the attachment mounts. Defaults to 0. */
-  threshold?: number | number[];
-  /** Getter for whether observation is enabled. Defaults to `() => true`. */
-  enabled?: () => boolean;
-};
+import type {
+  IntersectionAttachment,
+  IntersectionCallback,
+  UseIntersectionOptions,
+} from './use-intersection.types.ts';
 
 /**
  * Creates an attachment that observes the element with IntersectionObserver.
@@ -23,7 +14,7 @@ export type UseIntersectionOptions = {
 export function useIntersection(
   onIntersect: IntersectionCallback,
   options: UseIntersectionOptions = {},
-): Attachment<HTMLElement> {
+): IntersectionAttachment {
   const { root = null, rootMargin = '0px', threshold = 0, enabled = () => true } = options;
 
   return (node: HTMLElement) => {
