+++ packages/components/src/utilities/use-reduced-motion.types.ts
@@ -1,0 +1,4 @@
+export type UseReducedMotion = {
+  /** Reactive boolean — `true` when the user prefers reduced motion. */
+  readonly current: boolean;
+};
